Moved Tests from Android Tests (run on device) to Tests (run on machine)

+++ app/src/test/java/com/endorphinapps/kemikal/queenofclean/Entities/EmployeeTest.java
@@ -172,7 +172,7 @@ public class EmployeeTest {
 
     @Test
     public void getRateOfPay() throws Exception {
-        int rate = employee.getRateOfPay();
+        double rate = employee.getRateOfPay();
 
         assertEquals(123, rate);
     }
